flask-server/requirements.txt
alembic==1.5.8
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-editor==1.0.4
python-slugify==4.0.1
six==1.15.0
SQLAlchemy==1.4.2
text-unidecode==1.3
Werkzeug==1.0.1
